Change the config loader to be less strict.
Previously, we would raise a KeyError if there was a mismatch between
the config subclass and the config file. This meant that the user had to
configure the config.yaml to be identical to the BlackBox config class.

Instead of doing this, we'll now just return None or an empty list for
any parameters not configured. This has both upsides and downsides:

Advantage:
- The user now only needs to configure whatever they actually want to
use. No interest in notifiers? That's fine. Just configure the stuff you
want, and blackbox won't complain.

Downside:
- A poorly configured config.yaml will now fail silently. Less feedback
for the user. However, in my opinion, this is worth the tradeoff for
convenience. It's just so ugly if they have to create empty lists in the
 yaml.