Avoid storing file in memory during compression.
`file_path.read_bytes` stores the entire database backup in memory, and
the compressed result is also kept in memory, which increases memory
usage significantly if the database grows.

Therefore, we read the backup file in chunks, and gzip compress it on
the fly to a temporary file, which is removed after the sync finishes.