Make the upload directory input more flexible
The user shouldn't have to stress about the format
of their upload directory input.

Let's format it correctly before doing anything
with Google Drive. Do this by stripping leading
and trailing slashes, and replacing any duplicate
slashes // with single slashes /.

That would turn this:
/Hello//World///Hello/Foo/Bar
Into this:
Hello/World/Hello/Foo/Bar

With this pretty file path, we can then make sure
the file ends up in the correct place.

And, instead of creating duplicate folders like
we were before, we first _check_ if a folder with
the provided name and parent exists. If it does,
we use that one instead.