Implement support for rotation strategies using cron expressions
The user can include cron expressions under a
`rotation_strategies` option in `blackbox.yaml`.
These should be formatted as cron expressions,
with an additional sixth parameter representing
the number of backups matching this expression to
retain. If this number is 0, all backups matching
this expression will be removed. If it is "*",
then all backups matching this expression will be
retained.

I created a new abstract
method,`_delete_backup()`, that is used as a
helper to the new `_do_rotate()` method in the
storage handler base class. This will check the
given backup's date against the configured
rotation strategies and delete any backups that
should be deleted.

We continue to support the old `retention_days`
retention method as well. If `retention_days` is
configured, this will take precedence against the
`rotation_strategies` method.

Wrote tests to ensure the new cron/rotation
helpers work as expected.

So far, this has only been implemented for Google
Drive. The other handlers are broken because they
don't implement the new abstract method yet. That
will come next.